centralized the types interfaces
