+++ packages/undo-manager/src/types.ts
@@ -1,19 +1,65 @@
-export type HistoryChange = {
-	from: any;
-	to: any;
-};
+/**
+ * Represents a single change in history.
+ */
+export interface HistoryChange< T = unknown > {
+	/** The previous value */
+	from: T;
+	/** The new value */
+	to: T;
+}
 
-export type HistoryChanges = {
-	id: string | Record< string, any >;
-	changes: Record< string, HistoryChange >;
-};
+/**
+ * Represents changes for a single item.
+ */
+export interface HistoryChanges< T = unknown > {
+	/** The identifier for the item being changed */
+	id: string | Record< string, unknown >;
+	/** The changes made to the item */
+	changes: Record< string, HistoryChange< T > >;
+}
 
-export type HistoryRecord = Array< HistoryChanges >;
+/**
+ * Represents a record of history changes.
+ */
+export type HistoryRecord< T = unknown > = HistoryChanges< T >[];
 
-export type UndoManager = {
-	addRecord: ( record: HistoryRecord, isStaged: boolean ) => void;
-	undo: () => HistoryRecord | undefined;
-	redo: () => HistoryRecord | undefined;
+/**
+ * The undo manager interface.
+ */
+export interface UndoManager< T = unknown > {
+	/**
+	 * Record changes into the history.
+	 *
+	 * @param record   A record of changes to record.
+	 * @param isStaged Whether to immediately create an undo point or not.
+	 */
+	addRecord: ( record?: HistoryRecord< T >, isStaged?: boolean ) => void;
+
+	/**
+	 * Undo the last recorded changes.
+	 *
+	 * @return The undone record or undefined if nothing to undo.
+	 */
+	undo: () => HistoryRecord< T > | undefined;
+
+	/**
+	 * Redo the last undone changes.
+	 *
+	 * @return The redone record or undefined if nothing to redo.
+	 */
+	redo: () => HistoryRecord< T > | undefined;
+
+	/**
+	 * Check if there are changes that can be undone.
+	 *
+	 * @return Whether there are changes to undo.
+	 */
 	hasUndo: () => boolean;
+
+	/**
+	 * Check if there are changes that can be redone.
+	 *
+	 * @return Whether there are changes to redo.
+	 */
 	hasRedo: () => boolean;
-};
+}
